requirements.txt
Django~=4.2.3
djangorestframework~=3.14.0
psycopg2-binary==2.9.6
pytest-django==4.5.2
pytest-mock==3.11.1
django-environ==0.10.0

#development
pre-commit
black
flake8
isort
